cdk/requirements.txt
aws-cdk.core==1.134.0
aws-cdk.assertions==1.134.0
aws-cdk.aws-apigateway==1.134.0
aws-cdk.aws_s3==1.134.0
aws-cdk.aws_lambda==1.134.0
aws-cdk.aws_ssm==1.134.0
aws-cdk.aws-elasticbeanstalk==1.134.0
aws-cdk.aws-s3-assets==1.134.0
aws-cdk.aws-rds==1.134.0
aws-cdk.aws-ec2==1.134.0
aws-cdk.aws-ecs-patterns==1.134.0
aws-cdk.aws-ecs==1.134.0